WORK but fixing UNUSUAL TYPES

+++ handlers/summOfCallories.js
@@ -5,11 +5,16 @@ function summOfCallories(
 ) {
   if (portionFromSource) {
     if (/\d+(g|ml)/.test(portionFromSource)) {
-      portionFromSource = parseInt(portionFromSource);
+      portionFromSource = parseFloat(portionFromSource);
+      // console.log(caloriesForChosenPortion);
+      // console.log(dishPortionFromUserMessage);
+      // console.log()
+
       const summ = (
-        (caloriesForChosenPortion / parseInt(portionFromSource)) *
+        (caloriesForChosenPortion / parseFloat(portionFromSource)) *
         dishPortionFromUserMessage
       ).toFixed(0);
+      // console.log(parseFloat(portionFromSource));
       return summ;
     } else {
       console.log("vozvrat NULL tam gde CUP,MUG");
